Update readme.md
The readme could be more specific, as this is a laravel package, it might be better to update the example. I've added the use Facade, and changed Doctrine:: to EntityManager::, because that is what is declared in config/app.php  as a facade. Using the examples provided in the current readme are not working.